Move takeSamples out of SimulationResults and into GetSimulationResultsAction::getResourceSamples
