+++ examples/banananation/src/test/java/gov/nasa/jpl/aerie/banananation/Main.java
@@ -1,7 +1,6 @@
 package gov.nasa.jpl.aerie.banananation;
 
 import gov.nasa.jpl.aerie.merlin.driver.SerializedActivity;
-import gov.nasa.jpl.aerie.merlin.driver.SimulationResults;
 import gov.nasa.jpl.aerie.merlin.protocol.types.SerializedValue;
 import org.apache.commons.lang3.tuple.Pair;
 
@@ -39,6 +38,7 @@ public final class Main {
 
     final var simulationResults = SimulationUtility.simulate(schedule, simulationDuration);
 
-    System.out.println(SimulationResults.resourceSamples(simulationResults));
+    System.out.println(simulationResults.discreteProfiles);
+    System.out.println(simulationResults.realProfiles);
   }
 }
